requirements.txt
babel==2.16.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.9
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.2.2
Flask==3.0.3
ghp-import==2.1.0
griffe==1.5.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.2.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.48
mkdocs-material-extensions==1.3.1
mkdocstrings==0.27.0
mkdocstrings-python==1.12.2
mypy==1.13.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.0
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
Pygments==2.18.0
pymdown-extensions==10.12
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
ruff==0.8.3
six==1.17.0
tenacity==9.0.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.0.6
zipp==3.21.0
pillow==11.0.0
